+++ src/components/progressBar/ProgressBar.style.ts
