+++ src/components/title/subTitle/SubTitle.style.ts
@@ -3,5 +3,6 @@ import {StyleSheet} from 'react-native';
 export const subTitleStyles = StyleSheet.create({
   container: {
     paddingVertical: 8,
+    paddingHorizontal: 16,
   },
 });
