+++ src/navigators/StackNavigator.tsx
@@ -3,7 +3,7 @@ import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import LoginScreen from '@/screens/login/LoginScreen';
 import SignupScreen from '@/screens/signup/SignupScreen';
 import BottomTabNavigator from './bottomTabNavigator';
-import SignupCompleteScreen from '@/screens/signup/signupComplete/SIgnupCompleteScreen';
+import SignupCompleteScreen from '@/screens/signupComplete/SIgnupCompleteScreen';
 import {RootStackParamList} from '@/types/navigator';
 import MyPageScreen from '@/screens/myPage/MyPageScreen';
 
@@ -12,12 +12,12 @@ function StackNavigator() {
   return (
     <Stack.Navigator screenOptions={{headerShown: false}}>
       {/*로그인 페이지*/}
-      <Stack.Screen name={stackScreens.LoginScreen} component={LoginScreen} />
+      {/* <Stack.Screen name={stackScreens.LoginScreen} component={LoginScreen} />
       <Stack.Screen name={stackScreens.SignupScreen} component={SignupScreen} />
       <Stack.Screen
         name={stackScreens.SignupCompleteScreen}
         component={SignupCompleteScreen}
-      />
+      /> */}
       {/*BottomTab 3개 페이지*/}
       <Stack.Screen
         name={stackScreens.BottomTabScreens}
